refactor: remove side effects
