--- blink-core/src/main/kotlin/com.jordyma.blink/feed/domain/model/PromptResponse.kt
@@ -1,15 +1,0 @@
-package com.jordyma.blink.feed.domain.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class PromptResponse(
-
-    val subject: String,
-
-    val summary: String,
-
-    val keyword: List<String>,
-
-    val category: List<String>,
-)
