+++ blink-core/src/main/kotlin/com.jordyma.blink/feed/domain/service/FeedSummarizeService.kt
@@ -5,7 +5,7 @@ import com.jordyma.blink.feed.domain.Source
 
 interface FeedSummarizeService {
     fun updateSummarizedFeed(
-        content: SummaryContent,
+        content: PromptResponse,
         brunch: Source,
         feedId: Long,
         userId: Long,
